backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
gunicorn==21.2.0

torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

transformers==4.39.3
safetensors==0.4.3

Pillow==10.3.0
numpy==1.26.4
requests==2.32.2
scikit-learn==1.4.2
scipy==1.13.0
opencv-python==4.9.0.80
